refactor: setupScreen() 이동
