+++ app/src/main/java/woowacourse/movie/presenter/MoviePresenter.kt
@@ -1,0 +1,4 @@
+package woowacourse.movie.presenter
+
+class MoviePresenter {
+}
