+++ app/src/main/java/woowacourse/movie/ui/view/BaseActivity.kt
@@ -1,0 +1,22 @@
+package woowacourse.movie.ui.view
+
+import androidx.activity.enableEdgeToEdge
+import androidx.annotation.LayoutRes
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import woowacourse.movie.R
+
+abstract class BaseActivity : AppCompatActivity() {
+    fun setupScreen(
+        @LayoutRes layoutResId: Int,
+    ) {
+        enableEdgeToEdge()
+        setContentView(layoutResId)
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            insets
+        }
+    }
+}
