+++ app/src/main/java/woowacourse/movie/ui/view/movies/MoviesActivity.kt
@@ -1,21 +1,18 @@
-package woowacourse.movie.ui.view.main
+package woowacourse.movie.ui.view.movies
 
 import android.content.Intent
 import android.os.Bundle
 import android.widget.ListView
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import woowacourse.movie.R
-import woowacourse.movie.domain.model.MovieRepository
+import woowacourse.movie.domain.repository.MovieRepository
 import woowacourse.movie.ui.adapter.MovieAdapter
+import woowacourse.movie.ui.view.BaseActivity
 import woowacourse.movie.ui.view.booking.BookingActivity
 
-class MainActivity : AppCompatActivity() {
+class MoviesActivity : BaseActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setupScreen()
+        setupScreen(R.layout.activity_main)
         val adapter =
             MovieAdapter(
                 this,
@@ -29,15 +26,5 @@ class MainActivity : AppCompatActivity() {
             )
         val listView = findViewById<ListView>(R.id.movies)
         listView.adapter = adapter
-    }
-
-    private fun setupScreen() {
-        enableEdgeToEdge()
-        setContentView(R.layout.activity_main)
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
-            insets
-        }
     }
 }
